requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
boto3==1.42.9
botocore==1.42.9
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
jmespath==1.0.1
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
twilio==9.8.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
yarl==1.22.0
